requirements.txt
boto3
pandas
IPython
python-dotenv
google-cloud-bigquery
google-auth